pyproject.toml
[tool.poetry]
name = "geo-city-api"
version = "0.1.0"
description = ""
authors = ["irumako <irumakoTS@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.10.3"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
